+++ tests/src/Drupal10/Rector/Deprecation/ActionAnnotationToAttributeRector/config/configured_rule.php
@@ -12,5 +12,7 @@ return static function (RectorConfig $rectorConfig): void {
         new AnnotationToAttributeConfiguration('10.2.0', '12.0.0', 'Action', 'Drupal\Core\Action\Attribute\Action'),
         new AnnotationToAttributeConfiguration('10.2.0', '12.0.0', 'Block', 'Drupal\Core\Action\Attribute\Block'),
         new AnnotationToAttributeConfiguration('10.2.0', '12.0.0', 'ContentEntityType', 'Drupal\Core\Entity\Attribute\ContentEntityType'),
+        new AnnotationToAttributeConfiguration('10.2.0', '12.0.0', 'FormElement', 'Drupal\Core\Render\Attribute\FormElement'),
+        new AnnotationToAttributeConfiguration('10.2.0', '12.0.0', 'ViewsField', 'Drupal\views\Attribute\ViewsField'),
     ]);
 };